Handle memory pool in crio_next, fix crash on repeated filter fail
The pool mark is captured and reset after each filter evaluation
instead of only when a record passes a filter.  Otherwise, if many
records do not pass the filter, we exceed the allocated pool size.

This could use some further cleanup.  For testing, the non-pool
version of crio_next will create a new temporary pool.